Moved MD parsing to Data in preparation for use when parsing Issue Comments. Data may later be refactored.
